OBPIH-6108 Move Each price calculation into separate hook
